+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## v25.0.0-SNAPSHOT - under development
 
-TBD
+### 🎁 New Features
+
+* `Grid` accepts a `defaultCol` config that is applied to all columns unless the individual
+  column specifies otherwise.  (specificity wins)
+
 
 ## v24.1.1 - 2019-07-01